Refactored all tests related to current convert() built-in function
